requirements.txt
requests
nose
nose-testconfig